1. update nuget package 2. add karaoke skinnable column test scene
