+++ osu.Game.Rulesets.Karaoke.Tests/Skinning/TestSceneLyricLine.cs
@@ -12,12 +12,11 @@ using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Objects.Drawables;
 using osu.Game.Rulesets.Mods;
-using osu.Game.Tests.Visual;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Skinning
 {
     [TestFixture]
-    public class TestSceneLyricLine : SkinnableTestScene
+    public class TestSceneLyricLine : KaraokeSkinnableTestScene
     {
         public override IReadOnlyList<Type> RequiredTypes => new[]
         {
