+++ osu.Game.Rulesets.Karaoke.Tests/Skinning/TestSceneNotePlayfield.cs
@@ -7,7 +7,7 @@ using osu.Game.Rulesets.Karaoke.UI;
 
 namespace osu.Game.Rulesets.Karaoke.Tests.Skinning
 {
-    public class TestSceneNotePlayfield : KaraokeSkinnableTestScene
+    public class TestSceneNotePlayfield : KaraokeSkinnableColumnTestScene
     {
         [BackgroundDependencyLoader]
         private void load(RulesetConfigCache configCache)
